fix: restore update_company_fixtures call and backfill tds_section and entity_type in tax withholding categories
(cherry picked from commit f57e619ffd63eb489122442ff2a75ebf737060c3)
