+++ india_compliance/patches.txt
@@ -36,8 +36,8 @@ execute:from india_compliance.gst_india.setup import create_email_template; crea
 india_compliance.patches.post_install.update_payment_entry_fields
 india_compliance.patches.v15.remove_ignore_reconciliation_field
 india_compliance.patches.v15.migrate_print_options_to_new_field
-india_compliance.patches.post_install.update_company_fixtures #3
 india_compliance.patches.v14.migrate_and_update_tds_section_as_per_income_tax_act_2025
+india_compliance.patches.post_install.update_company_fixtures #3
 india_compliance.patches.post_install.update_itc_classification_field
 india_compliance.patches.v14.update_purchase_reco_email_template
 india_compliance.patches.v14.delete_purchase_receipt_standard_custom_fields
@@ -74,7 +74,7 @@ india_compliance.patches.v15.multiple_pi_in_boe
 india_compliance.patches.v15.update_supplier_filing_status
 india_compliance.patches.v14.set_pending_boe_qty #1
 india_compliance.patches.v15.remove_itc_amount_custom_fields
-india_compliance.patches.v14.update_tax_withholding_categories
+india_compliance.patches.v14.update_tax_withholding_categories #1
 india_compliance.patches.v16.remove_legacy_report_fixtures
 india_compliance.patches.v15.unset_auth_token
 india_compliance.patches.v15.migrate_logo_for_printing
