+++ india_compliance/patches/post_install/update_company_fixtures.py
@@ -1,13 +1,11 @@
 import frappe
 from erpnext.setup.setup_wizard.operations.taxes_setup import get_or_create_tax_group
-from frappe.query_builder.functions import IfNull
 
 from india_compliance.gst_india.overrides.company import (
     make_default_customs_accounts,
     make_default_gst_expense_accounts,
     make_default_tax_templates,
 )
-from india_compliance.income_tax_india.constants import OLD_TDS_SECTIONS, TDS_ENTITY_TYPE
 from india_compliance.income_tax_india.overrides.company import (
     create_company_fixtures as create_income_tax_fixtures,
 )
@@ -15,7 +13,6 @@ from india_compliance.income_tax_india.overrides.company import (
 
 def execute():
     company_list = frappe.get_all("Company", filters={"country": "India"}, pluck="name", order_by="lft asc")
-    set_section_and_entity_type_in_tax_withholding_category()  # execute before creating fixtures
 
     for company in company_list:
         # Income Tax fixtures
@@ -70,29 +67,3 @@ def update_root_for_rcm(company):
         output_account,
         update_modified=False,
     )
-
-
-def set_section_and_entity_type_in_tax_withholding_category():
-    doctype = frappe.qb.DocType("Tax Withholding Category")
-    categories = (
-        frappe.qb.from_(doctype)
-        .select(doctype.name)
-        .where(IfNull(doctype.tds_section, "") == "")
-        .where(IfNull(doctype.entity_type, "") == "")
-        .run(pluck=True)
-    )
-
-    for category_name in categories:
-        splitted_name = category_name.split(" - ")
-        # old naming [TDS - SECTION - ENTITY]
-        if len(splitted_name) < 3:
-            continue
-
-        if splitted_name[1] in OLD_TDS_SECTIONS and splitted_name[-1] in TDS_ENTITY_TYPE:
-            (
-                frappe.qb.update(doctype)
-                .set("tds_section", splitted_name[1])
-                .set("entity_type", splitted_name[-1])
-                .where(doctype.name == category_name)
-                .run()
-            )
